# uint256: optimize Lsh, Rsh, SRsh