@holiman 2024-05-06
numbers from my machine look good: 
```
[user@work uint256]$ benchstat bench.shifts.1 bench.shifts.2
name                    old time/op  new time/op  delta
Lsh/n_eq_0/uint256-8    3.91ns ±29%  3.93ns ±36%     ~     (p=0.670 n=10+10)
Lsh/n_gt_192/uint256-8  5.11ns ±32%  3.23ns ±41%  -36.78%  (p=0.001 n=10+10)
Lsh/n_gt_128/uint256-8  6.90ns ±31%  6.14ns ±31%     ~     (p=0.143 n=10+10)
Lsh/n_gt_64/uint256-8   9.43ns ±32%  6.23ns ±14%  -33.94%  (p=0.001 n=10+8)
Lsh/n_gt_0/uint256-8    12.1ns ±31%   8.1ns ±27%  -33.38%  (p=0.007 n=10+10)
Rsh/n_eq_0/uint256-8    3.85ns ±32%  4.80ns ± 0%  +24.93%  (p=0.000 n=10+8)
Rsh/n_gt_192/uint256-8  4.48ns ±33%  3.49ns ±36%     ~     (p=0.063 n=10+10)
Rsh/n_gt_128/uint256-8  5.23ns ±19%  6.29ns ±34%     ~     (p=0.515 n=8+10)
Rsh/n_gt_64/uint256-8   10.5ns ±36%   8.0ns ±29%  -23.47%  (p=0.011 n=10+10)
Rsh/n_gt_0/uint256-8    12.6ns ±40%  10.0ns ± 0%  -20.61%  (p=0.033 n=10+8)
```
Numbers aside, the code in this PR is a lot nicer than the existing!