@holiman 2024-05-06
> Why n_eq_0 regress in your testing? The behavior is the same.

Yeah, I don't know. Benchmarking is a bit black magic, getting good reproducible and "sane" results isn't always easy.